# src/swagger/openapi.yaml
openapi: 3.0.0
info:
  title: Attendance API
  version: 1.0.0
  description: API for managing student attendance and classes
servers:
  - url: http://localhost:3000
    description: Development server
components:
  securitySchemes:
    bearerAuth:
      type: apiKey
      in: header
      name: Authorization
      description: JWT token - provide token directly without "Bearer" prefix
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: object
    SuccessListResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: array
          items:
            type: object
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: string
paths:
...
  /class:
    post:
      summary: Create a new class
      tags: [Classes]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [className]
              properties:
                className:
                  type: string
                  minLength: 1
                  description: Name of the class (required, at least 1 character)
            example:
              className: "Advanced Mathematics 2024"
      responses:
        201:
          description: Class created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
              example:
                success: true
                data:
                  _id: "456def"
                  className: "Math 101"
                  teacherId: "123abc"
                  studentIds: []
        400:
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                success: false
                error: "Invalid request schema"
        401:
          description: Unauthorized - Invalid or missing token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                success: false
                error: "Unauthorized or invalid token"
        403:
          description: Forbidden - Only teachers can create classes
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                success: false
                error: "Forbidden, teacher access required"
        500:
          description: Server error creating class
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                success: false
                error: "Unknown error occurred"
  /class/{id}:
    get:
      summary: Get class details
      tags: [Classes]
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            pattern: '^[0-9a-fA-F]{24}$'
          description: Class ID (MongoDB ObjectId format - must be 24 hexadecimal characters)
      responses:
        200:
          description: Class retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
              example:
                success: true
                data:
                  _id: "456def"
                  className: "Math 101"
                  teacherId: "123abc"
                  studentIds: ["507f191e436d609404116113", "65a1b2c3d4e5f6789abcdef0"]
        401:
          description: Unauthorized - Invalid or missing token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                success: false
                error: "Unauthorized or invalid token"
        403:
          description: Forbidden - Teacher not class owner or student not enrolled
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                success: false
                error: "Forbidden, not class teacher"
        404:
          description: Class not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                success: false
                error: "Class not found"
        500:
          description: Server error retrieving class
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                success: false
                error: "Unknown error occurred"
  /class/{id}/add-student:
    post:
      summary: Add a student to a class
      tags: [Classes]
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            pattern: '^[0-9a-fA-F]{24}$'
          description: Class ID (MongoDB ObjectId format - must be 24 hexadecimal characters)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [studentId]
              properties:
                studentId:
                  type: string
                  pattern: '^[0-9a-fA-F]{24}$'
                  minLength: 1
                  description: Student ID to add to class (MongoDB ObjectId format - must be 24 hexadecimal characters)
            example:
              studentId: "507f1f77bcf86cd799439011"
      responses:
        200:
          description: Student added to class successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
              example:
                success: true
                data:
                  _id: "507f1f77bcf86cd799439011"
                  className: "Advanced Mathematics 2024"
                  teacherId: "507f191e436d609404116113"
                  studentIds: ["507f1f77bcf86cd799439012"]
        400:
          description: Validation error or user is not a student
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                success: false
                error: "Invalid request schema"
        401:
          description: Unauthorized - Invalid or missing token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                success: false
                error: "Unauthorized or invalid token"
        403:
          description: Forbidden - Not class teacher or user is not a student
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                success: false
                error: "Forbidden, not class teacher"
        404:
          description: Class or student not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                success: false
                error: "Class not found"
        409:
          description: Student already enrolled in class
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                success: false
                error: "Student already enrolled in class"
        500:
          description: Server error adding student
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                success: false
                error: "Unknown error occurred"
  /students:
    get:
      summary: Get all students
      tags: [Classes]
      security:
        - bearerAuth: []
      responses:
        200:
          description: Students retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessListResponse'
              example:
                success: true
                data:
                  - _id: "507f1f77bcf86cd799439011"
                    name: "John Doe"
                    email: "john@example.com"
                  - _id: "507f191e436d609404116113"
                    name: "Jane Smith"
                    email: "jane@example.com"
        401:
          description: Unauthorized - Invalid or missing token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                success: false
                error: "Unauthorized or invalid token"
        403:
          description: Forbidden - Only teachers can access this endpoint
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                success: false
                error: "Forbidden, teacher access required"
        500:
          description: Server error fetching students
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                success: false
                error: "Unknown error occurred"
...